Change imperial scale calculation to use int for DPI

+++ floorplan-to-structured-2d/floor_plan.py
@@ -50,7 +50,7 @@ class FloorPlan:
         if not DPI:
             DPI = self.hyperparameters["modelling"]["scale_adoption"]["dpi"]["in_use"]
         scale_on_paper_length = float(architectural_scale_normalized.split(':')[0].strip('`'))
-        imperial = float(DPI) * scale_on_paper_length
+        imperial = int(DPI) * scale_on_paper_length
         return (1 / imperial, 1 / imperial)
 
     @classmethod
